[fix] e-conomic book invoice button issue

+++ component/admin/views/order/tmpl/default.php
@@ -7,7 +7,8 @@
  * @license     GNU General Public License version 2 or later; see LICENSE
  */
 require_once JPATH_COMPONENT_SITE . '/helpers/product.php';
-$producthelper = new producthelper();
+
+$producthelper = new producthelper;
 
 require_once JPATH_COMPONENT_ADMINISTRATOR . '/helpers/order.php';
 require_once JPATH_COMPONENT_ADMINISTRATOR . '/helpers/shipping.php';
@@ -17,15 +18,14 @@ global $context;
 
 $app = JFactory::getApplication();
 
-$order_function = new order_functions();
-$config = new Redconfiguration();
-$option = JRequest::getVar('option');
+$order_function = new order_functions;
+$config = new Redconfiguration;
 $filter = $app->getUserStateFromRequest($context . 'filter', 'filter', 0);
 $lists = $this->lists;
 $model = $this->getModel('order');
-$redhelper = new redhelper();
-$shippinghelper = new shipping();
-$stockroomhelper = new rsstockroomhelper();
+$redhelper = new redhelper;
+$shippinghelper = new shipping;
+$stockroomhelper = new rsstockroomhelper;
 $dispatcher = JDispatcher::getInstance();
 JPluginHelper::importPlugin('redshop_product');
 
@@ -40,37 +40,48 @@ JPluginHelper::importPlugin('redshop_product');
 			form.task.value = pressbutton;
 		}
 
-		if (pressbutton == 'add') {
-			<?php      $link = 'index.php?option=' . $option . '&view=addorder_detail';
-			   $link = $redhelper->sslLink($link);
+		if (pressbutton == 'add')
+		{
+		<?php
+			$link = 'index.php?option=com_redshop&view=addorder_detail';
+			$link = $redhelper->sslLink($link);
 		?>
 			window.location = '<?php echo $link;?>';
 			return;
-			//form.view.value="addorder_detail";
 		}
-		if ((pressbutton == 'allstatus')) {
-			if (document.getElementById('order_status_all').value == '0') {
-				alert('<?php echo JText::_('COM_REDSHOP_SELECT_NEW_STATUS' ); ?>');
+
+		if ((pressbutton == 'allstatus'))
+		{
+			if (document.getElementById('order_status_all').value == '0')
+			{
+				alert('<?php echo JText::_('COM_REDSHOP_SELECT_NEW_STATUS'); ?>');
 				return false;
 			}
 
-		} else if ((pressbutton == 'edit') || (pressbutton == 'remove')) {
+		}
+		else if ((pressbutton == 'edit') || (pressbutton == 'remove'))
+		{
 			form.view.value = "order_detail";
-		} else if (pressbutton == 'multiprint_order') {
+		}
+		else if (pressbutton == 'multiprint_order')
+		{
 			form.view.value = "order";
 		}
 
-		try {
+		try
+		{
 			form.onsubmit();
 		}
-		catch (e) {
+		catch (e)
+		{
+			console.log(e);
 		}
 
 		form.submit();
 	}
 </script>
 
-<form action="<?php echo JRoute::_('index.php?option=' . $option . '&view=order'); ?>" method="post" name="adminForm"
+<form action="<?php echo JRoute::_('index.php?option=com_redshop&view=order'); ?>" method="post" name="adminForm"
       id="adminForm">
 <div id="editcell">
 <table class="adminlist" width="100%">
@@ -105,28 +116,35 @@ JPluginHelper::importPlugin('redshop_product');
 	<th class="title" width="10%">
 		<?php echo JHTML::_('grid.sort', 'COM_REDSHOP_ORDER_NUMBER', 'order_number', $this->lists['order_Dir'], $this->lists['order']); ?>
 	</th>
-	<?php if (ECONOMIC_INTEGRATION == 1 && ECONOMIC_INVOICE_DRAFT == 2)
-	{ ?>
+	<?php
+	if (ECONOMIC_INTEGRATION == 1 && ECONOMIC_INVOICE_DRAFT == 2)
+	{
+		?>
 		<th width="10%">
-			<?php echo  JHTML::_('grid.sort', 'COM_REDSHOP_BOOKINVOICE_NUMBER', 'bookinvoice_number', $this->lists['order_Dir'], $this->lists['order']); ?>
+			<?php echo JHTML::_('grid.sort', 'COM_REDSHOP_BOOKINVOICE_NUMBER', 'bookinvoice_number', $this->lists['order_Dir'], $this->lists['order']); ?>
 		</th>
-	<?php } ?>
+	<?php
+	}
+	?>
 	<th width="15%">
 		<?php echo JHTML::_('grid.sort', 'COM_REDSHOP_FULLNAME', 'uf.firstname', $this->lists['order_Dir'], $this->lists['order']); ?>
 	</th>
 	<th width="31%">
 		<?php echo JHTML::_('grid.sort', 'COM_REDSHOP_ORDER_STATUS', 'order_status', $this->lists['order_Dir'], $this->lists['order']); ?>
 	</th>
-	<?php if (USE_STOCKROOM == 1)
-	{ ?>
+	<?php
+	if (USE_STOCKROOM == 1)
+	{
+		?>
 		<th width="15%">
 			<?php echo JText::_('COM_REDSHOP_STOCKROOM_NAME'); ?>
 		</th>
 		<th width="10%">
 			<?php echo JText::_('COM_REDSHOP_STOCKROOM_DELIVERY_TIME'); ?>
 		</th>
-	<?php } ?>
-
+	<?php
+	}
+	?>
 	<th width="7%" nowrap="nowrap">
 		<?php echo JHTML::_('grid.sort', 'COM_REDSHOP_ORDER_DATE', 'cdate', $this->lists['order_Dir'], $this->lists['order']); ?>
 	</th>
@@ -135,62 +153,77 @@ JPluginHelper::importPlugin('redshop_product');
 	</th>
 	<th>
 	</th>
-	<?php if (POSTDK_INTEGRATION)
-		echo "<th></th>";?>
+	<?php
+	if (POSTDK_INTEGRATION)
+	{
+		echo "<th></th>";
+	}
+	?>
 </tr>
 </thead>
 <?php
 $send_mail_to_customer = 0;
+
 if (SEND_MAIL_TO_CUSTOMER)
 {
 	$send_mail_to_customer = "checked";
 }
 
 $k = 0;
+
 for ($i = 0, $n = count($this->orders); $i < $n; $i++)
 {
-	$row = & $this->orders[$i];
+	$row     = & $this->orders[$i];
 	$row->id = $row->order_id;
-	$link = 'index.php?option=' . $option . '&view=order_detail&task=edit&cid[]=' . $row->order_id;
-	$link = $redhelper->sslLink($link);
-	/**
+	$link    = 'index.php?option=com_redshop&view=order_detail&task=edit&cid[]=' . $row->order_id;
+	$link    = $redhelper->sslLink($link);
+
+	/*
 	 * @var $data
 	 * Trigger event onAfterDisplayProduct
 	 * Show content return by plugin directly into product page after display product title
 	 */
-	$data->highlight = new stdClass();
-	$results = $dispatcher->trigger('toHighlightGrid', array(& $row));
+	$data->highlight                  = new stdClass;
+	$results                          = $dispatcher->trigger('toHighlightGrid', array(& $row));
 	$data->highlight->toHighlightGrid = trim(implode("\n", $results));
 	?>
 	<tr class="<?php echo "row$k"; ?>">
 		<td class="order">
 			<?php echo $this->pagination->getRowOffset($i); ?>
-			<?php echo $data->highlight->toHighlightGrid;?>
+			<?php echo $data->highlight->toHighlightGrid; ?>
 		</td>
 		<td class="order">
 			<?php echo JHTML::_('grid.id', $i, $row->id); ?>
 		</td>
 		<td align="center">
-			<a href="<?php echo $link; ?>"
-			   title="<?php echo JText::_('COM_REDSHOP_EDIT_ORDER'); ?>"><?php echo $row->order_id; ?></a>
+			<?php
+				echo "<a href='$link' title='" . JText::_('COM_REDSHOP_EDIT_ORDER') . "'>$row->order_id</a>";
+			?>
 		</td>
 		<td align="center"><?php echo $row->order_number; ?></td>
 		<?php
 		if (ECONOMIC_INTEGRATION == 1 && ECONOMIC_INVOICE_DRAFT == 2 && $row->invoice_no && $row->is_booked == 1 && $row->bookinvoice_number)
 		{
-			?>
-			<td align="center"><?php echo $row->bookinvoice_number; ?></td>
-		<?php }?>
-		<td><?php
+			echo "<td align=\"center\">$row->bookinvoice_number</td>";
+		}
+		?>
+		<td>
+		<?php
+
 			echo $row->firstname . ' ' . $row->lastname;
-			echo ($row->is_company && $row->company_name != "") ? "<br />" . $row->company_name : ""; ?>
+
+			if ($row->is_company && $row->company_name != "")
+			{
+				echo "<br />" . $row->company_name;
+			}
+		?>
 		</td>
 		<td>
 			<table cellpadding="0" cellspacing="0" border="0">
 				<tr>
 					<td>
 						<?php
-						$linkupdate = JRoute::_('index.php?option=' . $option . '&view=order&task=update_status&return=order&order_id[]=' . $row->order_id);
+						$linkupdate = JRoute::_('index.php?option=com_redshop&view=order&task=update_status&return=order&order_id[]=' . $row->order_id);
 						echo $order_function->getstatuslist('order_status' . $row->order_id, $row->order_status, "class=\"inputbox\" size=\"1\" ");
 						echo "&nbsp";
 						echo $order_function->getpaymentstatuslist('order_paymentstatus' . $row->order_id, $row->order_payment_status, "class=\"inputbox\" size=\"1\" ");
@@ -198,23 +231,21 @@ for ($i = 0, $n = count($this->orders); $i < $n; $i++)
 					</td>
 				</tr>
 				<tr>
-					<td><textarea
-							name="customer_note<?php echo $row->order_id; ?>"><?php echo $row->customer_note;?></textarea>
+					<td>
+						<?php
+							echo "<textarea name='customer_note$row->order_id'>$row->customer_note</textarea>";
+						?>
 					</td>
-					<!--<td>
-							<input class="inputbox" name="ic<?php echo $row->order_id ;?>" id="include_comment<?php echo $row->order_id ;?>" value="N" type="checkbox" onclick="if(this.checked==true) {this.value = 'Y';} else {this.value = 'N';}"><?php echo JText::_('COM_REDSHOP_INCLUDE_COMMENT_MSG' ); ?><br />
-							<input class="inputbox" name="nc<?php echo $row->order_id ;?>" id="notify_customer<?php echo $row->order_id ;?>" value="N" type="checkbox" onclick="if(this.checked==true) {this.value = 'Y';} else {this.value = 'N';}"><?php echo JText::_('COM_REDSHOP_NOTIFY_CUSTOMER_MSG' ); ?>
-						</td>-->
 				</tr>
 				<tr>
 					<td>
-						<input type="checkbox" <?php echo $send_mail_to_customer;?>  value=""
-						       name="sendordermail<?php echo $row->order_id; ?>"
-						       id="sendordermail<?php echo $row->order_id; ?>"/><?php echo JText::_('COM_REDSHOP_SEND_ORDER_MAIL'); ?>
+						<?php
+							echo "<input type='checkbox' $send_mail_to_customer value='' name='sendordermail$row->order_id' id='sendordermail$row->order_id' />";
+							echo '<label for="sendordermail' . $row->order_id . '">' . JText::_('COM_REDSHOP_SEND_ORDER_MAIL') . '</label>';
+						?>
 					</td>
 				</tr>
 				<tr>
-					<!-- <td></td>-->
 					<td>
 						<input class="button"
 						       onclick="location.href = '<?php echo $linkupdate; ?>&status='+document.adminForm.order_status<?php echo $row->order_id; ?>.value+'&customer_note='+encodeURIComponent(document.adminForm.customer_note<?php echo $row->order_id; ?>.value)+'&order_sendordermail='+document.adminForm.sendordermail<?php echo $row->order_id; ?>.checked+'&order_paymentstatus='+document.adminForm.order_paymentstatus<?php echo $row->order_id; ?>.value  ; "
@@ -224,25 +255,28 @@ for ($i = 0, $n = count($this->orders); $i < $n; $i++)
 				</tr>
 			</table>
 		</td>
-		<?php if (USE_STOCKROOM == 1)
-		{ ?>
+		<?php
+		if (USE_STOCKROOM == 1)
+		{
+			?>
 			<td align="center">
-				<?php $order_items = $order_function->getOrderItemDetail($row->order_id);
+			<?php
+				$order_items = $order_function->getOrderItemDetail($row->order_id);
 
 				$stockroom_id = "";
+
 				for ($st = 0; $st < count($order_items); $st++)
 				{
 					if ($order_items[$st]->stockroom_id != "")
 					{
 						$stockroom_id .= $order_items[$st]->stockroom_id . ",";
 					}
-
 				}
 
 				if ($stockroom_id != "")
 				{
-
 					$stockroom_list = $stockroomhelper->getStockroom(substr_replace($stockroom_id, "", -1));
+
 					for ($s = 0; $s < count($stockroom_list); $s++)
 					{
 						echo $stockroom_list[$s]->stockroom_name;
@@ -251,58 +285,78 @@ for ($i = 0, $n = count($this->orders); $i < $n; $i++)
 						echo "<br>";
 					}
 				}
-
-				?>
-
+			?>
 			</td>
-			<td align="center"> <?php
+			<td align="center">
+			<?php
+
 				if ($stockroom_id != "")
 				{
-					$max_delivery = $stockroomhelper->getStockroom_maxdelivery(substr_replace($stockroom_id, "", -1));
-					$orderdate = $config->convertDateFormat($row->cdate);
+					$max_delivery  = $stockroomhelper->getStockroom_maxdelivery(substr_replace($stockroom_id, "", -1));
+					$orderdate     = $config->convertDateFormat($row->cdate);
 
-					$stamp = mktime(0, 0, 0, date('m', $row->cdate), date('d', $row->cdate) + $max_delivery[0]->max_del_time, date('Y', $row->cdate));
+					$stamp         = mktime(0, 0, 0, date('m', $row->cdate), date('d', $row->cdate) + $max_delivery[0]->max_del_time, date('Y', $row->cdate));
 					$delivery_date = date('d/m/Y', $stamp);
-					$current_date = date('d/m/Y');
-					$datediff = $stockroomhelper->getdateDiff($stamp, time());
+					$current_date  = date('d/m/Y');
+					$datediff      = $stockroomhelper->getdateDiff($stamp, time());
+
 					if ($datediff < 0)
 					{
 						$datediff = 0;
 					}
 
 					echo $datediff . " " . $max_delivery[0]->delivery_time;
-				} ?> </td>
-		<?php } ?>
+				}
+			?>
+			</td>
+	<?php
+		}
+	?>
 		<td align="center">
 			<?php echo $config->convertDateFormat($row->cdate); ?>
 		</td>
 		<td>
-			<?php echo $producthelper->getProductFormattedPrice($row->order_total);//,false,$row->order_item_currency);//CURRENCY_SYMBOL.number_format($row->order_total,2,PRICE_SEPERATOR,THOUSAND_SEPERATOR); ?>
+			<?php echo $producthelper->getProductFormattedPrice($row->order_total); ?>
 		</td>
-		<td><?php if ($row->invoice_no != '' && $row->is_booked == 0)
+		<td>
+			<?php
+
+			if ($row->invoice_no != '')
 			{
-				if ($row->is_company == 1 && $row->ean_number != "")
+				if ($row->is_booked == 0 && $row->bookinvoice_date <= 0)
 				{
-					echo JText::_('COM_REDSHOP_MANUALY_BOOK_INVOICE_FROM_ECONOMIC');
-				}
-				else
-				{
-					$confirm = 'if(confirm(\'' . JText::_('COM_REDSHOP_CONFIRM_BOOK_INVOICE') . '\')) { document.invoice.order_id.value=\'' . $row->order_id . '\';document.invoice.bookInvoiceDate.value=document.getElementById(\'bookDate' . $i . '\').value;document.invoice.submit(); }';
-					if ($row->order_payment_status == 'Paid' || $row->order_status == 'PR' || $row->order_status == 'C')
+					if ($row->is_company == 1 && $row->ean_number != "")
 					{
-						$confirm = 'document.invoice.order_id.value=\'' . $row->order_id . '\';document.invoice.bookInvoiceDate.value=document.getElementById(\'bookDate' . $i . '\').value;document.invoice.submit();';
+						echo JText::_('COM_REDSHOP_MANUALY_BOOK_INVOICE_FROM_ECONOMIC');
 					}
-					echo JHTML::_('calendar', date('Y-m-d'), 'bookDate' . $i, 'bookDate' . $i, $format = '%Y-%m-%d', array('class' => 'inputbox', 'size' => '15', 'maxlength' => '19'));    ?>
-					<input type="button" class="button" value="<?php echo JText::_("COM_REDSHOP_BOOK_INVOICE"); ?>"
-					       onclick="javascript:<?php echo $confirm; ?>"><br/>
-				<?php
+					else
+					{
+						$confirm = 'if(confirm(\'' . JText::_('COM_REDSHOP_CONFIRM_BOOK_INVOICE') . '\')) { document.invoice.order_id.value=\'' . $row->order_id . '\';document.invoice.bookInvoiceDate.value=document.getElementById(\'bookDate' . $i . '\').value;document.invoice.submit(); }';
+
+						if ($row->order_payment_status == 'Paid' || $row->order_status == 'PR' || $row->order_status == 'C')
+						{
+							$confirm = 'document.invoice.order_id.value=\'' . $row->order_id . '\';document.invoice.bookInvoiceDate.value=document.getElementById(\'bookDate' . $i . '\').value;document.invoice.submit();';
+						}
+
+						echo JHTML::_('calendar', date('Y-m-d'), 'bookDate' . $i, 'bookDate' . $i, $format = '%Y-%m-%d', array('class' => 'inputbox', 'size' => '15', 'maxlength' => '19'));    ?>
+						<input type="button" class="button" value="<?php echo JText::_("COM_REDSHOP_BOOK_INVOICE"); ?>"
+						       onclick="javascript:<?php echo $confirm; ?>"><br/>
+					<?php
+					}
+				}
+				elseif ($row->bookinvoice_date > 0)
+				{
+					echo JText::_('COM_REDSHOP_INVOICE_BOOKED_ON') . "<br />" . $config->convertDateFormat($row->bookinvoice_date);
 				}
 			}
+
 			echo "</td>";
 			$details = explode("|", $shippinghelper->decryptShipping(str_replace(" ", "+", $row->ship_method_id)));
+
 			if (strstr($details[0], 'default_shipping') && POSTDK_INTEGRATION)
 			{
 				echo "<td>";
+
 				if ($row->order_label_create)
 				{
 					echo JTEXT::_("COM_REDSHOP_XML_ALREADY_GENERATED");
@@ -324,7 +378,7 @@ for ($i = 0, $n = count($this->orders); $i < $n; $i++)
 ?>
 <tfoot>
 <td colspan="13">
-	<?php  echo $this->pagination->getListFooter(); ?>
+	<?php echo $this->pagination->getListFooter(); ?>
 </td>
 </tfoot>
 </table>
